Fixed property registration
I was getting correct behavior with incorrect code because of a variable I had stored in temporary memory. Once I restarted Blender it cleared out the temp variables, and I realized that I had failed to register the main "is the drawer open?" boolean.

It and the other booleans are now added in register, and deleted in unregister.