Added 3D View Panel
Besides the Text Editor, one of the other areas I thought would benefit most from this addon is the 3D Viewport.
Especially with the addition of the Asset Browser, I will want to access that, the File Browser, the Timeline, or Shader or Geometry Nodes without switching workspaces or sacrificing the screenspace of my 3d viewport.
It's not working as smoothly as the one in the Text Editor, I'm not sure why yet, but the 3D viewport seems to be catching clicks as click and drag or click and hold, and the menus are not disappearing immediately. Will require some investigation.